chore: extract template helpers to separate file
